+++ ErrorReporting.cpp
@@ -25,7 +25,7 @@ void ErrorReporting::setErrorCode(int errorCodeI, int errorCode){
 	errorCodes[errorCodeI] = errorCode;
 }
 
-void ErrorReporting::setNextErrorCode(int errorCode, double frequency){
+void ErrorReporting::setNextErrorCode(int errorCode, float frequency){
 	//int errorCodeIndex = 0;
 	//for(errorCodeIndex= 0; errorCodeIndex < MAX_ERROR_CODES; errorCodeIndex++){
 	//	if (errorCodes[errorCodeIndex] == errorCode){
